Re-expose `el` (#1230)
* re-expose `el`

We used to expose a custom `el` when we used a `setup()` and `render()`
function. But due to some refactors we got now only have a `setup()` and
no more `el`. This causes some problems if people relied on the exposed
`el`.

In this PR will make sure to re-expose the `el` that we used to expose.

The only issue is, now that we manually expose a list, we have to
re-expose the `$el` internal as well.

* update changelog